feat(api): login timeout
We're introducing a login timeouts in the API to handle repeated failed
login attempts with incorrect passwords for the same user.

The system allows users two attempts without blocking. If a third attempt
with an incorrect password occurs, the API will block further login
attempts for a duration of 1 minute. Subsequent failed attempts will
exponentially increase this timeout value by a factor of 4. The e quation
used to calculate the timeout is as follows:

    F(x) = c + min(4^(t - 3), M)

Attempts' persistence increases along with the timeout, being at least
1.5 times greater than the timeout and at most M. Attempts can never have
a duration less than 2 minutes. This means that an attempt must last for:

    F(y) = c + max(min((4^(a - 3) * 1.5), M), 2)

Where:
    - x is the timeout duration in minutes.
    - y is the attempt duration in minutes.
    - c is the current Unix timestamp in seconds.
    - a is the attempt number.
    - M is the maximum timeout value, specified by the
      "SHELLHUB_MAXIMUM_LOGIN_TIMEOUT" environment variable.

For example, if M equals 32768 (equivalent to 15 days), the timeouts for
different attempt numbers are as follows:

               timeout(x) | attempt(y)
    - F(3)  = [ c + 1,       c + 2     ]
    - F(4)  = [ c + 4,       c + 6     ]
    - F(5)  = [ c + 16,      c + 24    ]
    - F(8)  = [ c + 1024,    c + 1536  ]
    - F(11) = [ c + 32768,   c + 32768 ]

To achieve this functionality, each failed login attempt will increase
a `login-attempt` count by 1, which persists for M minutes and is reset
if a successful attempt is made. After the third attempt, the system will
save a `login-timeout`, which is the current timestamp plus the timeout
value.

Attempts persist for at least timeout * 1.5 and at most M; after that,
the attempt count is reset, which means that new failed attempts will
count from the beginning. The timeout will persist for a maximum of M
minutes; after that, the timeout is reset.

The M value is controlled with the "MAXIMUM_LOGIN_TIMEOUT" environment
variable. When it equals 0, this feature is disabled. The default value
is 32768, representing 15 days.

Consumers must inspect the `X-Timeout` header, which will indicate the
end of the timeout in UTC seconds. When no timeout is set, the header
value will be 0.